# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: HEXA Purchase Assistant API
  version: 0.1.0
  description: |
    FastAPI backend for Flutter app, WhatsApp webhook, and Super Admin.
    AI is used for parsing only; business logic stays server-side.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8000
    description: Local
...
paths:
...
  /v1/businesses/{businessId}/analytics/summary:
    get:
      tags: [analytics]
      summary: Overview KPIs for date range
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsSummary'
  /v1/businesses/{businessId}/analytics/insights:
    get:
      tags: [analytics]
      summary: Home insights (top item, alerts)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Insights payload
  /v1/businesses/{businessId}/analytics/categories:
    get:
      tags: [analytics]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Category rows
  /v1/businesses/{businessId}/analytics/suppliers:
    get:
      tags: [analytics]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Supplier rows
  /v1/businesses/{businessId}/analytics/brokers:
    get:
      tags: [analytics]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Broker rows
  /v1/businesses/{businessId}/analytics/items:
    get:
      tags: [analytics]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Item breakdown rows
...
  /v1/admin/businesses:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Tenant list
  /v1/admin/api-usage:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Usage by provider
  /v1/admin/feature-flags:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Global feature flags
  /v1/admin/integrations:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Integration configuration status
  /v1/admin/audit-logs:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Audit log placeholder
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AnalyticsSummary:
      type: object
      properties:
        total_purchase:
          type: number
        total_qty_base:
          type: number
        total_profit:
          type: number
        purchase_count:
          type: integer
        base_unit_label:
          type: string
          example: kg
...